rag_gemini_chatbot_app/rag_gemini_chatbot_app/requirements.txt

streamlit
langchain
langchain-google-genai
sentence-transformers
faiss-cpu
